More type fixes
