+++ app/types/model.ts
@@ -1,0 +1,10 @@
+import type { ModelInfo } from '~/utils/types';
+
+export type ProviderInfo = {
+  staticModels: ModelInfo[],
+  name: string,
+  getDynamicModels?: () => Promise<ModelInfo[]>,
+  getApiKeyLink?: string,
+  labelForGetApiKey?: string,
+  icon?:string,
+};
